ingestion/requirements.txt
pyyaml>=6.0
gitpython>=3.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
markdown>=3.4.0
tqdm>=4.66.0

